refactors test aggregate by into multiple modules
